builder: Fix deprecated regect warning
Warning: $* is no longer supported
In perl the ($*) would force the regex to match all lines in a string
this is useful when your string is not just a single line, but may have
newline character in it.  This was replaced by /m swith and disabled
after 5.10.

This fix removes that warning.

Also added a .gitignore